requirements.txt
# Python 3.10.17
# torch==2.5.0, torchvision==0.20
#python_version == 3.10.17
flask
opencv-python
pupil-apriltags
ultralytics
torch==2.5.0
torchvision==0.20
coremltools
robotpy-wpimath
scipy
numpy
matplotlib 